# docs/openapi.yaml
openapi: 3.0.3

info:
  title: Power Outage Reporting API
  version: 1.0.0
  description: |
    API for submitting power outage reports.

    This API uses simulated data only and does not connect to real EVN systems.
  contact:
    name: Power Outage Reporting Project

servers:
  - url: http://localhost:8000/api/v1
    description: Local Kong Gateway

tags:
  - name: Outage Reports
    description: Operations for reporting power outages.

paths:
  /outage-reports:
    post:
      tags:
        - Outage Reports
      summary: Submit a power outage report
      description: |
        Validates and stores a new power outage report. After the report is
        stored, the service publishes an `outage.reported` event to RabbitMQ.
      operationId: createOutageReport
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/CorrelationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOutageReportRequest"
            examples:
              validRequest:
                summary: Valid power outage report
                value:
                  customerCode: CUST00001
                  servicePointCode: SP00001
                  reporterName: Nguyen Van A
                  phoneNumber: "0901234567"
                  address: 123 Tran Thai Tong, Hanoi
                  description: Power outage in the entire house
      responses:
        "201":
          description: The outage report was created successfully.
          headers:
            Location:
              description: URL of the created outage report.
              schema:
                type: string
                example: /api/v1/outage-reports/ee26f857-220c-46e7-8c99-f66ea9a041ef
            X-Correlation-Id:
              $ref: "#/components/headers/CorrelationId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutageReportResponse"
              examples:
                created:
                  value:
                    id: ee26f857-220c-46e7-8c99-f66ea9a041ef
                    reportCode: OUT-20260801-00001
                    status: RECEIVED
                    message: The power outage report has been received.
                    createdAt: "2026-08-01T09:30:00+07:00"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"

    get:
      tags: [Outage Reports]
      summary: List outage reports
      description: Returns reports in reverse creation order. Restricted to OPERATOR and ADMIN.
      operationId: listOutageReports
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [RECEIVED]
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: A page of outage reports.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutageReportPageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /outage-reports/{reportId}:
    get:
      tags: [Outage Reports]
      summary: Get an outage report
      operationId: getOutageReport
      security:
        - bearerAuth: []
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The requested outage report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutageReportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by Keycloak.

  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        A unique UUID generated by the client. Reusing the same key with the
        same request returns the previously created result.
      schema:
        type: string
        format: uuid
      example: c1203d7d-a58f-45ea-b9ec-469d77b24871
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      description: Request identifier used for tracing across services.
      schema:
        type: string
        format: uuid
      example: a7a5a3ee-1a43-40aa-a285-ed13bc822612

  headers:
    CorrelationId:
      description: Request identifier used for tracing across services.
      schema:
        type: string
        format: uuid

  schemas:
    CreateOutageReportRequest:
      type: object
      additionalProperties: false
      required:
        - customerCode
        - servicePointCode
        - reporterName
        - phoneNumber
        - address
      properties:
        customerCode:
          type: string
          minLength: 1
          maxLength: 30
          pattern: "^[A-Z0-9-]+$"
          example: CUST00001
        servicePointCode:
          type: string
          minLength: 1
          maxLength: 30
          pattern: "^[A-Z0-9-]+$"
          example: SP00001
        reporterName:
          type: string
          minLength: 1
          maxLength: 200
          example: Nguyen Van A
        phoneNumber:
          type: string
          pattern: "^(0|\\+84)[0-9]{9}$"
          example: "0901234567"
        address:
          type: string
          minLength: 1
          maxLength: 500
          example: 123 Tran Thai Tong, Hanoi
        description:
          type: string
          maxLength: 1000
          example: Power outage in the entire house

    OutageReportResponse:
      type: object
      required:
        - id
        - reportCode
        - status
        - message
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        reportCode:
          type: string
          example: OUT-20260801-00001
        status:
          type: string
          enum:
            - RECEIVED
          example: RECEIVED
        message:
          type: string
          example: The power outage report has been received.
        createdAt:
          type: string
          format: date-time
          example: "2026-08-01T09:30:00+07:00"

    OutageReportPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OutageReportResponse"
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer

    ValidationErrorDetail:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: phoneNumber
        message:
          type: string
          example: Phone number has an invalid format.

    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - code
        - message
        - path
        - correlationId
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-08-01T09:30:00+07:00"
        status:
          type: integer
          format: int32
          example: 400
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: The request contains invalid data.
        details:
          type: array
          items:
            $ref: "#/components/schemas/ValidationErrorDetail"
        path:
          type: string
          example: /api/v1/outage-reports
        correlationId:
          type: string
          format: uuid
          example: a7a5a3ee-1a43-40aa-a285-ed13bc822612

  responses:
    BadRequest:
      description: The request contains invalid data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2026-08-01T09:30:00+07:00"
            status: 400
            code: INVALID_REQUEST
            message: The request contains invalid data.
            details:
              - field: phoneNumber
                message: Phone number has an invalid format.
            path: /api/v1/outage-reports
            correlationId: a7a5a3ee-1a43-40aa-a285-ed13bc822612
    Unauthorized:
      description: The access token is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated user does not have the required role.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The idempotency key was reused with different request data.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested outage report does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"